+++ include/type.h
@@ -45,9 +45,7 @@ m_bool isprim(Type type);
 m_bool isa(Type var, Type parent);
 m_bool verify_array(Array_Sub array);
 Type new_array_type(Env env, m_uint depth, Type base_type, Nspc owner_nspc);
-static inline Type find_common_anc(Type lhs, Type rhs) {
-  return isa(lhs, rhs) > 0 ? rhs : isa(rhs, lhs) > 0 ? lhs : NULL;
-}
+Type find_common_anc(Type lhs, Type rhs);
 m_uint id_list_len(ID_List list);
 void type_path(m_str str, ID_List path);
 m_bool env_add_value(Env env, m_str name, Type type, m_bool is_const,  void* value);
